requirements.txt
fastapi==0.115.5
python-dotenv==1.0.1
python-telegram-bot==21.7
uvicorn==0.32.1
